+++ libs/angular-remote-components/src/lib/model/remote-component.ts
@@ -1,0 +1,5 @@
+import { RemoteComponentConfig } from './remote-component-config.model'
+
+export interface ocxInitRemoteComponent {
+  ocxRemoteComponent(config: RemoteComponentConfig): void
+}